Upgrade to DataMapper 0.9.2.
First sweep of the code to get most of the tests passing.  The tests that are
failing are likely due to an odd behavior in dm-core that is documented in
test/dm_core_test.rb.  Some features (possibly all development related) were
lost in the porting process.  Hopefully, after resolving all the TODOs, the
lost features will be reclaimed.  Overall, the app seems to be working.